draft: utilization on ip pool detail
